feat: add swagger and folder refactoring
