+++ src/main/java/com/wowmate/server/chatroom/domain/Chatroom.java
@@ -1,0 +1,27 @@
+package com.wowmate.server.chatroom.domain;
+
+import com.wowmate.server.BaseEntity;
+import lombok.Getter;
+import lombok.Setter;
+
+import javax.persistence.*;
+
+@Entity
+@Getter
+@Setter
+public class Chatroom extends BaseEntity {
+
+    @Id
+    @GeneratedValue
+    @Column(name = "chatroom_id")
+    private Long id;
+
+    @OneToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "create_chatroom_id")
+    private CreateChatroom createChatroom;
+
+    private Long postId;
+    private Long writeUserId;
+    private Long replyUserId;
+
+}
