+++ src/main/java/com/wowmate/server/chatroom/domain/Message.java
@@ -1,0 +1,26 @@
+package com.wowmate.server.chatroom.domain;
+
+import com.wowmate.server.BaseEntity;
+import lombok.Getter;
+import lombok.Setter;
+
+import javax.persistence.*;
+
+@Entity
+@Getter
+@Setter
+public class Message extends BaseEntity {
+
+    @Id
+    @GeneratedValue
+    @Column(name = "message_id")
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "chatroom_id")
+    private Chatroom chatRoom;
+
+    private Long sendUserId;
+    private String content;
+
+}
