+++ src/main/java/com/wowmate/server/chatroom/repository/ChatroomRepository.java
@@ -1,0 +1,8 @@
+package com.wowmate.server.chatroom.repository;
+
+import com.wowmate.server.chatroom.domain.Chatroom;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ChatroomRepository extends JpaRepository<Chatroom, Long> {
+
+}
