+++ src/main/java/com/wowmate/server/chatroom/repository/MessageRepository.java
@@ -1,0 +1,7 @@
+package com.wowmate.server.chatroom.repository;
+
+import com.wowmate.server.chatroom.domain.Message;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MessageRepository extends JpaRepository<Message, Long>  {
+}
