+++ src/main/java/com/wowmate/server/comment/domain/Comment.java
@@ -1,0 +1,40 @@
+package com.wowmate.server.comment.domain;
+
+import com.wowmate.server.BaseEntity;
+import com.wowmate.server.post.domain.Post;
+import com.wowmate.server.user.domain.User;
+import lombok.Getter;
+import lombok.Setter;
+import lombok.ToString;
+
+import javax.persistence.*;
+import java.util.ArrayList;
+import java.util.List;
+
+@Entity
+@Getter
+@Setter
+@ToString
+public class Comment extends BaseEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    @Column(name = "comment_id")
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY, cascade = CascadeType.ALL)
+    @JoinColumn(name = "post_id")
+    private Post post;
+
+    @ManyToOne(fetch = FetchType.LAZY, cascade = CascadeType.ALL)
+    @JoinColumn(name = "user_id")
+    private User user;
+
+    @OneToMany(mappedBy = "comment", cascade = CascadeType.ALL)
+    private List<CommentReply> commentReplyList = new ArrayList<>();
+
+    private String content;
+
+    private int likeNumber;
+
+}
