+++ src/main/java/com/wowmate/server/comment/repository/CommentReplyRepository.java
@@ -1,0 +1,7 @@
+package com.wowmate.server.comment.repository;
+
+import com.wowmate.server.comment.domain.CommentReply;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CommentReplyRepository extends JpaRepository<CommentReply, Long> {
+}
