+++ src/main/java/com/wowmate/server/comment/repository/CommentRepository.java
@@ -1,0 +1,7 @@
+package com.wowmate.server.comment.repository;
+
+import com.wowmate.server.comment.domain.Comment;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CommentRepository extends JpaRepository<Comment, Long> {
+}
