+++ src/main/java/com/wowmate/server/post/domain/Category.java
@@ -1,0 +1,25 @@
+package com.wowmate.server.post.domain;
+
+import com.wowmate.server.BaseEntity;
+import lombok.Getter;
+import lombok.Setter;
+
+import javax.persistence.*;
+import java.util.ArrayList;
+import java.util.List;
+
+@Entity
+@Getter @Setter
+public class Category extends BaseEntity {
+
+    @Id @GeneratedValue
+    @Column(name = "category_id")
+    private Long id;
+
+    private String name;
+
+    @OneToMany(mappedBy = "category", cascade = CascadeType.ALL)
+    private List<PostCategory> postCategoryList = new ArrayList<>();
+
+
+}
