+++ src/main/java/com/wowmate/server/post/repository/CategoryRepository.java
@@ -1,0 +1,8 @@
+package com.wowmate.server.post.repository;
+
+import com.wowmate.server.post.domain.Category;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CategoryRepository extends JpaRepository<Category, Long> {
+
+}
