+++ src/main/java/com/wowmate/server/post/repository/PostCategoryRepository.java
@@ -1,0 +1,7 @@
+package com.wowmate.server.post.repository;
+
+import com.wowmate.server.post.domain.PostCategory;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PostCategoryRepository extends JpaRepository<PostCategory, Long> {
+}
