+++ src/main/java/com/wowmate/server/post/repository/PostRepository.java
@@ -1,0 +1,8 @@
+package com.wowmate.server.post.repository;
+
+import com.wowmate.server.post.domain.Post;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PostRepository extends JpaRepository<Post, Long> {
+
+}
