+++ src/main/java/com/wowmate/server/user/domain/Gender.java
@@ -1,0 +1,5 @@
+package com.wowmate.server.user.domain;
+
+public enum Gender {
+    M, F
+}
