--- src/main/java/server/WOWMATE/Server/test.java
@@ -1,6 +1,0 @@
-package server.WOWMATE.Server;
-
-public class test {
-
-
-}
