--- src/main/java/server/chatroom/repository/CreateChatroomRepositroy.java
@@ -1,7 +1,0 @@
-package server.chatroom.repository;
-
-import server.chatroom.domain.CreateChatroom;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CreateChatroomRepositroy extends JpaRepository<CreateChatroom, Long> {
-}
