--- src/main/java/server/comment/domain/Comment.java
@@ -1,40 +1,0 @@
-package server.comment.domain;
-
-import lombok.Getter;
-import lombok.Setter;
-import lombok.ToString;
-import server.BaseEntity;
-import server.post.domain.Post;
-import server.user.domain.User;
-
-import javax.persistence.*;
-import java.util.ArrayList;
-import java.util.List;
-
-@Entity
-@Getter
-@Setter
-@ToString
-public class Comment extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    @Column(name = "comment_id")
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY, cascade = CascadeType.ALL)
-    @JoinColumn(name = "post_id")
-    private Post post;
-
-    @ManyToOne(fetch = FetchType.LAZY, cascade = CascadeType.ALL)
-    @JoinColumn(name = "user_id")
-    private User user;
-
-    @OneToMany(mappedBy = "comment", cascade = CascadeType.ALL)
-    private List<CommentReply> commentReplyList = new ArrayList<>();
-
-    private String content;
-
-    private int likeNumber;
-
-}
