--- src/main/java/server/comment/repository/CommentReplyRepository.java
@@ -1,7 +1,0 @@
-package server.comment.repository;
-
-import server.comment.domain.CommentReply;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CommentReplyRepository extends JpaRepository<CommentReply, Long> {
-}
