--- src/main/java/server/post/repository/PostRepository.java
@@ -1,8 +1,0 @@
-package server.post.repository;
-
-import server.post.domain.Post;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PostRepository extends JpaRepository<Post, Long> {
-
-}
