--- src/main/java/server/user/domain/Gender.java
@@ -1,5 +1,0 @@
-package server.user.domain;
-
-public enum Gender {
-    M, F
-}
