--- src/main/java/server/user/repository/UserRepository.java
@@ -1,9 +1,0 @@
-package server.user.repository;
-
-import server.user.domain.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-
-    User findByEmail(String email);
-}
