--- src/test/java/WOWMATE/Server/ServerApplicationTests.java
@@ -1,13 +1,0 @@
-package WOWMATE.Server;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ServerApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
