+++ src/test/java/com/wowmate/server/ServerApplicationTests.java
@@ -1,0 +1,13 @@
+package com.wowmate.server;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ServerApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
